fix: Ensure global hash table exists no matter the join type
Previously this was only ensuring the hash table existed on a partition if the
join type was a LEFT join _and_ there was no global left visit bitmaps (meaning
no batches received from the right side of the join for any partition).

This was wrong because it's possible for a probe-side (right) partition to not
have received a batch, but other partitions might have. This would lead to the
branch being take where the join type is LEFT and there exists a global left
visit bitmap. And this branch did not set hash table reference for the
partition, leading to the panic during pull.

While we only need the hash table to exist for left joins, this change just
ensures the hash table exists no matter the join type or left visit state so
that I don't have to think too hard again.